Added function to skip already created barcodes, included counter at end

+++ Barcode/Barcode.py
@@ -1,6 +1,9 @@
 import barcode
+import datetime
 from barcode import Code128
 from barcode.writer import ImageWriter
+import os.path
+from os import path
 import tkinter as tk
 from tkinter import Tk
 from tkinter import messagebox
@@ -12,7 +15,9 @@ saveDirectory = ''
 
 def __main__():
     pathToFile = getFilePath()
+    print('Beginning processing...')
     generateBarcodes()
+    print('Program complete"')
 
 
 def getFilePath():
@@ -25,25 +30,38 @@ def getFilePath():
 
 
 def generateBarcodes():
+    processed = 0
     global pathToFile, saveDirectory
     errorList = []
     with open(pathToFile) as csv:
+        #record[0] = name
+        #record[1] = store
+        #record[2] = region
        for item in csv:
            record = item.split(',')
-           barcodeNumber = record[1].rstrip().zfill(15) #for a nice chunky barcode
-           fileName = record[1].rstrip() + ' ' + record[0]
-           barcodeText = record[0] + '\n' + record[1].rstrip()
+           barcodeNumber = record[0].strip() + record[2].zfill(15) + '-' + record[1].strip()
+           barcodeText = record[0].strip() + ' ' + record[2].strip() + '-' + record[1].strip()
+           fileName = record[2].strip() + '-' + record[1].strip() + ' ' + record[0].strip()
            try:
-               with open(saveDirectory + '/' + fileName + '.png', 'wb') as barcodeFile:
-                   Code128(barcodeNumber, writer=ImageWriter()).write(barcodeFile, text=barcodeText)                   
+               #If the barcode already exists in the folder, we won't waste time reprocessing it
+               if(path.exists(saveDirectory + '/' + fileName + '.png')):
+                   print('Barcode already generated for ' + fileName)
+               else:
+                   with open(saveDirectory + '/' + fileName + '.png', 'wb') as barcodeFile:
+                       Code128(barcodeNumber, writer=ImageWriter()).write(barcodeFile, text=barcodeText) 
+                       processed = processed + 1
            except:
                print('Error processing ' + fileName)
                errorList.append(record)
-    with open(saveDirectory + '/errors.txt', 'a') as errorLog:
-        for record in errorList:
-            for part in record:
-                errorLog.write(part)
-            errorLog.write('\n')
+       print('Processing complete. Processed ' + str(processed))
+    if errorList:
+        print('Writing error log...')
+        with open(saveDirectory + '/errors.txt', 'a') as errorLog:
+            errorLog.write('-----ERROR LOG-----\n' + str(datetime.datetime.now()) + '\n--------------\n')
+            for record in errorList:
+                for part in record:
+                    errorLog.write(part)
+                errorLog.write('\n')
 
 
 __main__()
